% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/response_curve.R
\name{response_curve}
\alias{response_curve}
\title{生成物种响应曲线}
\usage{
response_curve(
  file_paths,
  species_name,
  variable_name,
  x_limit = c(-25, 25),
  y_limit = c(0, 0.8),
  save_path = NULL,
  dpi = 600
)
}
\arguments{
\item{file_paths}{包含数据文件的文件路径向量}

\item{species_name}{物种名称（用于标题）}

\item{variable_name}{环境变量名称（用于标题和x轴标签）}

\item{x_limit}{x轴范围（默认为c(-25,25)）}

\item{y_limit}{y轴范围（默认为c(0,0.8)）}

\item{save_path}{保存图片路径（可选）}

\item{dpi}{图片分辨率（默认为600）}
}
\value{
ggplot对象
}
\description{
该函数读取多个数据文件，计算平均值和绝对偏差，并创建物种对环境变量的响应曲线。
}
